pyproject.toml
[tool.poetry]
name = "llama-api"
version = "0.1.0"
description = ""
authors = ["Roman Solomatin <36135455+Samoed@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
llama-cpp-python = "^0.1.83"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
pyupgrade = "^3.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# https://github.com/tiangolo/sqlmodel/blob/main/pyproject.toml
[tool.mypy]
python_version = "3.10"
# --strict
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
plugins = [
    "pydantic.mypy",
]
ignore_missing_imports = false
# --strict end
exclude = [
    "tests/"
]

[[tool.mypy.overrides]]
module = [
    "sqlalchemy_utils",
    "nltk",
    "pandas",
    "schedule",
    "uvicorn",
]
ignore_missing_imports = true

[tool.black]
line-length = 120
target-version = ['py311']
preview = true

[tool.coverage.report]
omit = [
    "main.py",
    "view_main.py",
    "__init__.py",
]
skip_empty = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::UserWarning",
]
